+++ backend/src/main/java/com/woowacourse/kkogkkog/auth/application/dto/MemberCreateResponse.java
@@ -1,4 +1,4 @@
-package com.woowacourse.kkogkkog.application.dto;
+package com.woowacourse.kkogkkog.auth.application.dto;
 
 import lombok.AccessLevel;
 import lombok.Getter;
