+++ backend/src/main/java/com/woowacourse/kkogkkog/auth/presentation/dto/InstallSlackAppRequest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.kkogkkog.presentation.dto;
+package com.woowacourse.kkogkkog.auth.presentation.dto;
 
 import lombok.AccessLevel;
 import lombok.Getter;
